Add regex to match file extension in image caching
The current way that the pathname in caching images was problematic, because it was impossible to add multiple file extensions on the endsWith string method. This makes it way easier to just include other file extensions. As I wrote in the comment above the changed line, you can  add in any other file extensions by including the necessary file extension name in the regex, ex. (jpg|png|gif).